refactor: transformed input masks into utils functions
